Restructuring of examples folder for consistency
