+++ neurobench/examples/gsc/model_data/s2s_gsc_snntorch
