+++ neurobench/examples/gsc/training.py
@@ -1,0 +1,1 @@
+# this file should contain the training file of the model
