+++ neurobench/examples/mackey_glass/echo_state_network.py
