+++ neurobench/examples/mswc_fscil/training.py
@@ -1,0 +1,3 @@
+# this file should containt the training code used for the benchmark
+def train(net, data):
+    pass
