+++ neurobench/examples/obj_detection/obj_det_model.py
@@ -1,0 +1,9 @@
+from neurobench.models.model import NeuroBenchModel
+
+## Define model ##
+class ObjDetectionModel(NeuroBenchModel):
+    def __init__(self, net, box_coder, head):
+        ...
+
+    def __call__(self, x):
+    	...
