+++ neurobench/examples/primate_reaching/model_data/model_parameters
@@ -1,0 +1,1 @@
+This folder should contain trained model parameters, if applicable.
