+++ neurobench/examples/primate_reaching/training.py
@@ -1,0 +1,1 @@
+# this file should contain the code used for training the model used for the benchmark
